docs: explicitly require difft --color=always command
Problem: I had trouble getting highlights to show up in the diff buffer.
I only later discovered that I had to force difft to use colors in the
output. It was not obvious, even though I set `GIT_EXTERNAL_DIFF=difft`
initially in the `command` in the `setup` method.

Solution: Make it clear in the README to use `difft --color=always` as
the `GIT_EXTERNAL_DIFF` environment variable.